Template: Add footer component
https://phabricator.endlessm.com/T31510
